+++ src/com/github/unchama/seichiassist/breakeffect/BlizzardTaskRunnable.java
@@ -13,7 +13,7 @@ import org.bukkit.scheduler.BukkitRunnable;
 import com.github.unchama.seichiassist.ActiveSkill;
 import com.github.unchama.seichiassist.data.Coordinate;
 import com.github.unchama.seichiassist.data.PlayerData;
-import com.github.unchama.seichiassist.util.Util;
+import com.github.unchama.seichiassist.util.BreakUtil;
 
 public class BlizzardTaskRunnable extends BukkitRunnable{
 	Player player;
@@ -45,12 +45,12 @@ public class BlizzardTaskRunnable extends BukkitRunnable{
 
 		if(playerdata.activeskilldata.skillnum > 2){
 			for(Block b : breaklist){
-				Util.BreakBlock(player, b, droploc, tool, false);
+				BreakUtil.BreakBlock(player, b, droploc, tool, false);
 				b.setType(Material.PACKED_ICE);
 			}
 		}else{
 			for(Block b : breaklist){
-				Util.BreakBlock(player, b, droploc, tool, true);
+				BreakUtil.BreakBlock(player, b, droploc, tool, true);
 				playerdata.activeskilldata.blocklist.remove(b);
 			}
 		}
