+++ src/com/github/unchama/seichiassist/breakeffect/VladmiaTaskRunnable.java
@@ -16,7 +16,7 @@ import com.github.unchama.seichiassist.SeichiAssist;
 import com.github.unchama.seichiassist.data.Coordinate;
 import com.github.unchama.seichiassist.data.PlayerData;
 import com.github.unchama.seichiassist.task.moveParticleTaskRunnable;
-import com.github.unchama.seichiassist.util.Util;
+import com.github.unchama.seichiassist.util.BreakUtil;
 
 public class VladmiaTaskRunnable extends BukkitRunnable{
 	SeichiAssist plugin = SeichiAssist.plugin;
@@ -63,7 +63,7 @@ public class VladmiaTaskRunnable extends BukkitRunnable{
 
 
 		for(Block b : breaklist){
-			Util.BreakBlock(player, b, droploc, tool, false);
+			BreakUtil.BreakBlock(player, b, droploc, tool, false);
 			new moveParticleTaskRunnable(player,b,Color.RED).runTaskTimer(plugin, 0, 3);
 		}
 	}
